requirements.txt
amqp==5.3.1
asgiref==3.8.1
beautifulsoup4==4.12.3
billiard==4.2.1
black==24.10.0
boto3==1.35.63
botocore==1.35.63
build==1.2.2.post1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.16
django-bootstrap-v5==1.0.11
django-bootstrap5==24.3
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-redis==5.4.0
django-storages==1.14.4
djangorestframework==3.15.2
djongo==1.3.7
flake8==7.1.1
gunicorn==23.0.0
iniconfig==2.0.0
jmespath==1.0.1
kombu==5.4.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.0.0
pip-tools==7.4.1
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
pymongo==3.11.4
pyproject_hooks==1.2.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
redis==5.2.0
s3transfer==0.10.3
setuptools==75.5.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.0
whitenoise==6.8.2
